+++ src/Service/CaseHelper.php
@@ -3,6 +3,8 @@
 namespace App\Service;
 
 use App\Entity\CaseEntity;
+use App\Entity\FenceReviewCase;
+use App\Entity\RentBoardCase;
 use App\Entity\ResidentComplaintBoardCase;
 use App\Repository\CasePartyRelationRepository;
 
@@ -26,7 +28,9 @@ class CaseHelper
     public function getRelevantTemplateAndPartiesByCase(CaseEntity $case): array
     {
         switch (get_class($case)) {
+            case RentBoardCase::class:
             case ResidentComplaintBoardCase::class:
+            case FenceReviewCase::class:
                 // Get relations from both sides
                 $complainantRelations = $this->relationRepository
                     ->findBy(['case' => $case, 'type' => ['Tenant', 'Representative'], 'softDeleted' => false]);